Much hacking and exploratory programming.
This is a checkpoint before I further explore the discovery I made that
just setting status.loadBalancer.ingress[0].ip is sufficient for dataplane
routing to work correctly, so there's no need to touch externalIPs.

This opens up some questions, notably whether the assigned-ip annotation
is needed at all. I'm thinking no, because the status field is protected
such that users aren't able to (easily) edit it, so we can use it as the
one truth in the system for assigned IPs.

It also opens up the question of whether we should mess with external
traffic policy at all, or let the user shoot themselves in the foot
as desired.