Check in the vendor directory. Fixes #20
I couldn't decide which one to do, so I asked twitter. The results
were pretty evenly split, with "check in vendor/" winning by a tiny
margin.

The fact that both options were ~equally voted for tells me that it
really doesn't matter and we just need to pick one. I said I would
do whichever one won on Twitter, so here we are.