Lock down the speaker pod. #53
The speaker pod is more privileged, because of ARP: it needs access
to the host networking stack, and it needs CAP_NET_RAW to receive and
transmit ARP.

Annoyingly, because of kubernetes/kubernetes#56374, we can't run the
container as non-root but still grant it net_raw. So the closest we
can get to minimum necessary privileges is to run as root, but drop
all caps except net_raw, and disallow privilege escalation so that
the container cannot regain other privileges.