"lock down" the test-bgp-router pod. #53
"Lock down" is a bit of a joke in this case. The pod needs sys_admin
and net_admin to tweak netfilter rules, so it's basically game over
for security anyway. We can't even make the filesystem RO because we
write configs at runtime.

Oh well, it's a testing deployment that should never ever be in prod,
so it should be okay.