Update vendoring yet again.
k8s, gobgp and cobra play really badly with each other, because of the
merry dependency hell that the `dep` crew are attempting to bolt onto
go. Because everything worked fine before, so obviously we should ruin
it by making me spend two hours getting a damn binary to compile by pinning
packages to versions from a year ago because one library is too lazy to update.

That makes complete sense. It sure is better than the world where the social
contract was that "world HEAD" should mostly build most of the time and you
can just freeze mostly any time and have things work right. This thing where
I have to solve NP-hard problems on behalf of computers because people are too
lazy to be a bit careful about API compatibility is clearly better!

This is all sarcasm. I hate it. I wish k8s would lead the way and loudly and
publicly abandon and shame `dep` and its friends.