Use hostname as the lock identity, and tweak timers some more.
The names of the timers sadly make zero sense. The retry timer is actually the
interval between writes to the lock, and the renew deadline is an internal
implementation detail that should never be exposed in the API. Whatever,
it works now, with a 10s lease timeout and a renewal every 5s.