Delete AUTHORS.
I wasn't doing a good job of keeping it up to date vs. the git commit
log, and Google's open-source guidelines only suggest using it to track
copyright ownership - which nobody has really asked me to do yet because
the project doesn't warrant that degree of fanciness.

So, let's just point to the commit log and the release notes credits, for
now at least.