Hide glog's flags from the commandline. #189
We are fortunate, in that we only register flags in main.go, so we can
just overwrite the base commandline flag set with a new one, which
effectively hides all flags that weren't defined in main(). Just before
doing that, we set --logtostderr=true for glog, so that it doesn't try
to do file I/O other than writing to os.Stderr.

This is step one of trying to remove glog from MetalLB.