Fix the controller mutation test to pass Service.Status between runs.
This more realistically reflects the desired test case of "the user is
doing a bunch of changes to the spec one after the other", because in
those cases we expect that the status from the previous mutation will
still be there.

Unfortunately this seems to have uncovered a bunch of bugs, so this
commit breaks the build. I'm committing it first so that I can
work on fixing each failing case in separate followup commits.