Never mind, I read my tests wrong.
Or rather: the test table isn't currently formulated to work with
the new setup I wanted. It does make sense as a setup, but I'm not
going to rewrite my entire test table now.

I can just change the test for pool changes to test the specific
regression I need.