Make usableNodes work correctly when a small number of pods for a deployment from multiple replicasets are on the same node when one is ready and the other is not
Closes danderson/metallb#474

The usableNodes function has a comment:
// usableNodes returns all nodes that have at least one fully ready
// endpoint on them.

The code removed pretty much makes it do the opposite of that. If
an endpoint has an address that isn't ready, a node will be set false even
if a different endpoint has an address that's ready. In a situation where
a deployment has two replicas, a rolling update causes the current
replicaset to scale down from 2 to 1. The new replicaset then scales
from 0 to 1. If the pod created for the new replicaset ends up on the
same physical node as the remaining pod from the old replicaset, this
code will mean metallb will not listen on your service IPs until both
pods enter a ready state or until the last remaining pod in the old
replicaset is terminated and the new pod is created on a different node.