website: Fix style in 0.9.2 release notes
This commit fixes the style so the paragraph is not dedented and the
next item is not shown in the same line.

The first issue is "fixed" by just making it one paragraph, as no other
alternatives were found that worked correctly in hugo 0.57.0 (didn't
explored ugly hacks like manually indent, etc.).

The second issue (item in the wrong line) was reproduced locally with
hugo 0.57.0. It didn't happen with other newer hugo versions (tested
several >= 0.62.2 versions), however this patch is rendered correcly on
new and old hugo versions.

This patch was tested with hugo 0.57.0, as this was the only version I
found to render the site locally as similar as possible to the current
website. In particular, it renders the site with the horrible black
thing rendered when written like this[1], and properly includes the site
header logo (is not shown with newer hugo versions). However, some other
changes might be required if the production site is rendered using
another hugo version.

Fixes: #548

[1]: https://github.com/metallb/metallb/commit/2d0cbd89bd71298f55372c168365acec711dc3f2#diff-3c4a3af03bacf924a0523da3eb3ec9bdR18