website: Fix typo to get contributors list
We should get the git log output only for the commits since the last
release to CUR-BRANCH (that can be main or v0.9, for example, if minor
or patch is release is created).

When searching the common ancestor (using git merge-base) we look for it
between CUR-BRANCH and PREV-TAG. When we retrive the logs, we want the
logs from the common ancestor to CUR-BRANCH too, that is why we searched
for the common ancestor in the first place.