Add source-interface configuration to BGP peer configuration
Add an optional "source-interface" parameter to the BGP peer
configuration.

When configured, this must be an interface name, and the BGP
session to the peer is sourced from that interface. This allows
the BGP session to originate from a Dummy interface (equivalent
to a Loopback interface on other BGP routers) rather than the
physical interface facing the BGP peer. This, in turn, can
help avoid conflicts when other processes (eg. Calico) are also
BGP peering off the same node.