Expose status endpoint
With the introduction of peer autodiscovery, BGP peers can now be
configured using two sources: static configuration and Node objects.
This fact makes it more difficult to understand why a speaker behaves
a certain way. Therefore, add a queryable status endpoint which
allows viewing the in-memory configuration of a speaker.

Although I haven't encountered race conditions in practice, it looks
like there could be a race if the status endpoint is queried during
an update to the peer list. Therefore, protect the peers slice with a
mutex.

The following changes are done in order to facilitate a clean
shutdown:

- Move cleanup goroutine to end of main function. It's better to
  concentrate all cleanup actions in one place.
- Stop SpeakerList inside cleanup goroutine rather than in a defer
  statement. Deferring causes ordering problems in this case.
- Add a WaitGroup to ensure proper termination of the various
  goroutines.

Signed-off-by: Johannes Liebermann <johannes@kinvolk.io>