Replace ForceSync() with resyncSvcCh chan
This commit prepares the removal of k8s.Client from SpeakerList.
Before this commit we had a circular dependency between
the k8s package and the speakerlist package.
k8s depends on controller, controller depends on speakerlist
for layer2, and speakerlist depends on k8s for ForceSync().
To break this loop, instead of calling k8s.ForceSync() from
speakerlist, we create a resyncSvcCh chan that we pass to
both k8s and speakerlist.

Signed-off-by: Etienne Champetier <echampetier@anevia.com>