website: Update maintainers description
I'm very new to the MetalLB community, but based on my limited
experience so far, it seemed like this text wasn't quite up to date
with the current state since I've seen other maintainers handling some
PRs.

This change also links to the contributing page as a place to find out
more info on who the current maintainers are.  That information is not
there currently, but it seems like an appropriate place to add it.