ci: Set up build and unit test CI jobs
This change sets up two CI jobs using Github actions.  This
environment seems like a good fit since it's free, included with
github, and capable of running e2e CI jobs based on kind.

In addition to the CI jobs, this change adds a new "unit" target for
invoke so that it's easy to manually run the same commands that are
used in CI.