Out of date insall instructions for kustomize
I'm not sure if referring to a URL was ever permitted in the resources section, but it is not possible in the kustomize version bundled with recent releases of `kubectl`. Instead, remote `kustomization.yaml` files can be referenced in the `bases:` section. I've also updated the tag to the current release, I'm not sure if there is a better way to ensure that the tag remains up to date.