layer2: Make nodes re-join memberlist cluster
Currently, nodes join the memberlist cluster only on startup and if they
are removed (e.g. small network downtime and they recover) they never
join again. This means the node is alive, but not taken into account as
a possible speaker, making it pointless to run the MetalLB speaker.
Furthermore, it can cause split brain and you might have multiple nodes
responding to ARP requests causing the TCP connection to be reset often
and have downtime in practice.

This commit adds a go routine to do memberlist joins in a loop. This was
missing in the commit that introduced memberlist
(3c5fc630a62a386c6955a506ab6c529bcf9c4069 "Use hashicorp/memberlist to
speedup dead node detection") and fixes issue #589.

This routine tries to join nodes in two situations: (a) every 1 minute,
(b) when the current k8s node object changes and the SetNode() callback
is called.

The case (b) is used when the node loses connectivity for long enough
for Kubernetes to mark it in a Not Ready state. When the node is Ready
again (e.g network is restored for the node), the callback is invoked
and the node tries to join the cluster.

The case (a), periodic checks, is needed for downtimes that are smaller
than the time Kubernetes takes to mark a node as Not Ready (about 45 sec
by default). For example, if a node has a small downtime (~15 seconds)
memberlist will remove it from the cluster but as this is short enough
for Kubernetes to not change the node state to Not Ready, the SetNode()
callback won't be invoked (as the node didn't change). In these cases,
the periodic join every 1 minute will make those node join the
memberlist cluster again.

The internal kubernetes watcher already has this logic to notify when
the current node changes. In the Layer2 case, though, it was just a
"return nil". Now, it queues the message to process a re-join.

This routine tries to join only the nodes that are not currently part of
the cluster, as it was advised by Etienne Champetier that the memberlist
Join() operation is not a no-op for nodes that are already members and
can take a considerable amount of time (specially when only a small
group of nodes need to re-join the cluster).

Furthermore, the IPs to join are stored in the SpeakerList struct to be
able to recover nodes even if the API server is down. This means we
don't need to query the API server and just use the IPs cached in the
struct. This was also very good advise from Etienne.

There is another go routine introduced here to update the IPs to join
stored in the struct. This go routine updates them by quering the API
server, every 5 minutes only to not overload the API server (something
very easy to do by mistake). This is in a separate go routine because,
as I mention in the code, the current version of client-go we are using
doesn't support a context or a timeout so that query can block. To avoid
this blocking other operations, we use a separate go routine. When we
update to new client-go versions, this will of course not be needed and
we can just run with a timeout if we want to.

Fixes: #589
Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>
(cherry picked from commit 9dd124f12c828fda8ebf5e197bdb78d67f4aa7c1)