Document maintainership of MetalLB
When I first joined the MetalLB development community, I was curious
who the current maintainers were.  This patch is a proposed method of
communicating who that team is, as well as how the team is governed.

First, create a CODEOWNERS file.  This is a github recognized file
format for documenting maintainers of all or parts of a repository.

This patch also introduces a document in the community section of the
web site that discusses the group of maintainers and how people are
added or removed.