bgp: Decode hold timer expired notification
While setting up a dev and test environment for BGP mode, I got the
following error in my speaker log:

{"caller":"bgp.go:408","error":"got BGP notification code 0x0400 (unknown code)",
    "event":"peerNotification","localASN":64513,
    "msg":"peer sent notification, closing session",
    "peer":"172.18.0.5:179","peerASN":64512,"ts":"2020-07-22T21:20:07.673523481Z"}

0x0400 is an unspecific "Hold timer expired" notification.  The code
portion of "04" is "Hold timer expired", and the sub-code of "00"
means unspecific, or no additional detail.

Update the code to provide an improved error message in this case.
The updated message is:

{"caller":"bgp.go:408",
    "error":"got BGP notification code 0x0400 (Hold Timer Expired (unspecific))",
    "event":"peerNotification","localASN":64513,
    "msg":"peer sent notification, closing session",
    "peer":"172.18.0.5:179","peerASN":64512,"ts":"2020-07-22T22:02:56.887511947Z"}

The related IANA assignments can be found here:
https://www.iana.org/assignments/bgp-parameters/bgp-parameters.xhtml#bgp-parameters-3