dev-env: Add BGP support
Expand the dev-env setup with an option to configure MetalLB in BGP
mode.  It will run frr in a container along side the kind cluster and
each of the speakers will peer with it.

Running `inv dev-env` will stand up a cluster with MetalLB
unconfigured as before.  If you run `inv dev-env -p bgp`, the BGP
configuration will be applied.

To keep the environment completely isolated, no routing is done
outside of the frr container.  You can get a shell in the container to
test connectivity to LoadBalancer IPs.  You can also use the vtysh
console of the router to see the router's status information.  A
README.md file is included in dev-env/bgp/ to demonstrate how to use
this bgp dev env.