dev-env: Allow interactive sudo
The use of sudo() does not allow the user to type in a password.  The
alternative of run("sudo ...") works fine, so use that instead.

In passing, clean up the tool output some more by adding a newline and
echoing a couple more commands to make it more clear what is
happening.