dev-env: Create vtysh.conf
Prior to this change, every time you ran the `vtysh` command in the
`frr` container when using the bgp dev environment, you would get an
error about `vtysh.conf` missing.  The error is misleading and makes
it seem like something is wrong.

Create an empty config file to suppress this error message.

Closes #713