Drop e2etest directory
These tests no longer work.  They rely on the now archived virtuakube,
which doesn't build against kube 1.19, which ends up blocking fixing
metallb for kube 1.19.  See #723 for discussion of where it gets in the
way.

At this point we should just drop this directory and re-introduce
e2etest code once someone is able to work on it.  We can always pull
this back out of git history if it helps.

The commit contents are the result of roughly:

 - git rm -r e2etest
 - go mod tidy
 - git add go.mod go.sum

Closes #192
Closes #629
Closes #639