layer2: Send GARPs on cluster membership changes
There have been multiple patches looking at alternatives for addressing
this issue, most recently #709.

Speaker ownership of IPs is recalculated each time membership of the
cluster changes. When a new speaker is chosen, that speaker will send
a set of gratuitous ARP or NDP messages to inform the network of the
new location for that IP address.

A problem existed where the cluster could end up in an inconsistent
state if a speaker lost ownership of an IP but didn't realize it.
In the meantime, a new speaker will have taken over the IP. When the
original owner rejoins the cluster, it will fail to reclaim the IP
with gratuitous ARP / NDP if it thinks it never lost it in the first
place.

We already force a resync of all services any time we detect a cluster
membership change via memberlist.  The only time this would result in a
new GARP being sent out is if the node saw this as previously not owning
the IP, but now owning.

This patch is a very simple option for improving this situation.
Instead of only sending a GARP the first time we see an IP in use, send
it every time we resync the Service.

The tradeoff taken with this approach is to increase the consistency of
the network at the cost of more freqently sending GARPs to the network.
It seems possible that there is such thing as "too much" and that some
sort of rate limiting should be done, though that is done here.

Closes #584