Consolidate CI jobs for publishing images
The only difference between the jobs which publish the controller and
the worker images is the --binaries argument. This argument can be
specified multiple times so we don't really need separate CI jobs for
publishing images.

One downside of this change is that the CI run time could slightly
increase, however I think it's negligible and I'd rather reduce the
amount of code we have to maintain.

Name the job "publish-images" instead of "deploy-{controller,speaker}".
"Deploy" is confusing since it sounds like we are deploying MetalLB
on some platform which isn't the case here.