website: Check for outdated dependencies when creating a release
It happened in the past that client-go was out of date and that broke
MetalLB on new Kubernetes releases, like seen in this PR:
	https://github.com/metallb/metallb/pull/812

When we created the 0.9 release branch no one really checked, we could
have updated it by then and never have this issue in the first place.
So, I propose to add this to the release-process.