Add operator to toleration
The default toleration doesn't specify an operator, so the operator defaults to Equal. It also doesn't specify a value, so there's nothing for Equal to match. The result is that the speaker Pods are never scheduled to controlplane nodes and never announce services located there (like Traefik in K3s).

The solution is to add operator: Exists to the tolerations key:
```
tolerations:
- effect: NoSchedule
  key: node-role.kubernetes.io/master
  operator: Exists
```
This is supported by the docs:

> The default value for operator is Equal.
>
> A toleration "matches" a taint if the keys are the same and the effects are the same, and:
>
> the operator is Exists (in which case no value should be specified), or
> the operator is Equal and the values are equal.

Demonstration of the failure is at [13:04 in my latest video](https://www.youtube.com/watch?v=9PLw1xalcYA&t=784s).

Fixes #814