Verify if interface master is up before skipping
In some cases such as wen using Open vSwitch, it is possible that
the interface master will be in down state (ovs-system) but MetalLB
skips the port because it is already assigned a master which means
that no ARP response ever goes through it.

This patch instead makes sure that the master interface is actually
up before skipping it to make sure that we can send traffic across
in one of the interfaces.

Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>