frr: Add initial FRR configuration templating engine
This commit implements the MetalLB BGP interface by generating
an FRR configuration file. It does this by maintaining an
in-memory representation of the MetalLB BGP configuration that
can be used to template the FRR configuration file. At each call
of the MetalLB BGP interface, the in-memory state is
regenerated and the config file is (re)created.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>