BGP: Add option to configure BGP speaker implementation type
In order to configure the speaker application to use different
BGP implementation types, add an option to allow the user to specify
the implementation type on speaker startup.

This option is set using an environment variable. It is set to "native"
by default. As there are currently no other BGP implementations, setting
this to another value will cause the speaker to not start.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>