FRR: Add frr-reloader directory and basic manifests for MetalLB with FRR
The frr-reloader will act as a proxy between the MetalLB speaker and the FRR sidecar container.
The speaker will render a valid FRR configuration file and place it as frr.conf in the volume it
shares with the reloader (/etc/frr_reloader). Then it would ask the reloader to load the given file
by sending it a SIGHUP. The speaker can know the reloader's PID by reading /etc/frr_reloader/reloader.pid.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>