Extend BGP test to check routes and session state
Currently, BGP test only verifies that the remote peer can hit the
service. We don't have guarantee that all the ip is advertised from all
the nodes and that the BGP sessions are established.

Here we add helper packages to check both, providing a base for adding
more testcases.

The frr parsing utilities are in a package of their own, and may be
useful if parsing the FRR output is needed to provide status feedback or
metrics.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>