Add Operator mode for e2e tests
Currently how the e2e tests work is by changing MetalLB's ConfigMap directly.
Here we add a way for the tests to use the Operator's CRDs instead of
manipulating the ConfigMap. That'll enable us to have a true e2e scenario
for the Operator, meaning that from one side (the Operator's e2e) it
validates that it configures MetalLB's ConfigMap the way it thinks it should,
and with the new tests mode we could also know that the Operator is
configuring MetalLB properly from the other side (MetalLBs e2e) = the way
MetalLB expects it to be configured.

This may also serve as a base for when(if) the Operator is integrated with
MetalLB and not as a standalone.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>