Add support to Podman-based kind environments
This is half-baked commit adding support to Podman-based kind
environments. It continues assuming the container runtime is docker but
checks behind the scenes if docker is a wrapper for podman, i.e. a
symlink to podman for cases where commands and/or outputs differ.

Users on Red Hat and Debian family distributions can either install
package 'podman-docker' or create a symbolic link docker -> podman.

For more information of how user:group ownership is mapped, please refer
to https://www.redhat.com/sysadmin/rootless-podman-makes-sense

Signed-off-by: Carlos Goncalves <cgoncalves@redhat.com>