bpg: Introduce SessionManager
In order to remove some package-scoped variables from the BGP FRR
implementation, this commit adds a SessionManager. The SessionManager
can be used to manage multiple BGP Sessions. The SessionManager is
created on creation of the BGP controller and is specific to the
BGP implementation. Currently, only the FRR BGP implementation
uses the SessionManager. In order for the BGP controller to create
BGP sessions, it must do so through the SessionManager.

The interface looks something like this:

sessionManager := newBgp("frr")
session := sessionManager.NewSession(...)
session.Set()
session.Close()

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>