allow CIDR range to be a single IP
The introduced AddressPool validation check for invalid ranges where the first address of the range is after the second, had a bug that disallows an address range with exactly one IP address. 
But a Single IP CIDR is a valid configuration.
Fixed the bug & extended a test case with a single IP configuration.

Furthermore, fixed a bug in the unit tests for address pool CIDRs, missing the protocol definition and therefore not working as intended.

Signed-off-by: Pumba98 <mail@pumba98.de>