Buffer FRR reloading commands
A change in config is reflected in several changes (and reloads) of the
config map. Here, decouple the request for a reload to the reloading.
If a reload request comes in, we squash all the next reload requests and
apply the final configuration only after 500 milliseconds.
If a new request comes after a reload, we start a new waiting period.

500 msec is short enough not to affect the user experience, and at the
same time is long enough to collect all the variations related to a
single config change.

We move the reloadConfig initialization over the main test function to
avoid the data race detector to complain about having two different
goroutines accessing it.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>