Add BFD support (#967)
* Enable BGP + BFD support

Here we enable an optional BFD profile that can be associated with a BGP
session. Using BFD in conjunction with BGP allows faster failover
detection.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Decouple the configuration structure of external containers to speakers

Filling routerid / holdtime in the external containers but using them as
configuration of speakers is misleading. Here we decouple it.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* E2ETests: cover BFD

Extend e2e tests to make sure bfd pairing works with the external
containers.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Skip BFD tests in CI in legacy lanes

BFD is FRR only so we won't cover it.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>