Fail when trying to load frr only features in native mode
Ideally, the checks of FRR only fields should happen in the config
handler of the speaker, which is implementation aware.
This is not possible to achieve because of the defaulting mechanism in
place.
Since it makes sense to have all the checks in the same place, and those
checks must happen before the defaulting, we inject an implementation
driven hook to check the raw config.

Current checks are meant to drop:
- ipv6 pools in case of native bgp mode
- ipv6 aggregation length
- BFD profiles

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>