Fix L2 metrics tests
Getting a lot of flakes in the CI:
```
      Expected
          <*errors.errorString | 0xc000695790>: {
              s: "metric metallb_layer2_requests_received not found",
          }
      to be nil
```
It is caused by the fact that this metric represents the amount
of ARP/NS that arrived to the speaker for the svc's ip, while the
test doesn't necessarily does one. That can happen because the speaker
sends a gratuitous message when he announces an IP.

Also added a check for the gratuitous_sent metric to cover that.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>