BGP Tests: use ipfamily specific pools
Now that we have a check that drops the config in case of native
implementation, if the address pools contain both v4 and v6 addresses,
the native bgp tests will fail. Here, we extend the tests to accept a
parameter that is the address range of the pool used for the test.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>