E2E: Add an empty bgpd.conf file with zebra password
Aims to fix a flake where BGP is not able to start, showing messages
as

BGP: [EC 33554499] sendmsg_nexthop: zclient_send_message() failed

the current-config is as well empty, so it could be an issue related to
the order in which bgp starts / the tests place a config under its hood.
Here we start with a valid config that gets eventually overridden by the
tests.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>