E2E Test: provide support for dual stack
Instead of driving the ipfamily from outside, we use the family of the
loadbalancer ip to derive what family it is, and we use both in case of
dual stack services.
We add more entries to the test tables forcing the service to be dual
stack in case of dual stack entries, and adding extra entries to check
single v6 in case of dual stack (for the v4, we defer to the fact that
a service gets created as single stack v4 in case of dual stack).

Additionally, the old ipFamily parameter is now used to choose the
ipfamily of the pairing between the external FRRs and the speakers,
making it possible to test scenarios where we advertise v6 addresses on
a v4 connection and viceversa.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>