Let speakers settle in L2 ExternalTrafficPolicy=Local tests
Since we're creating multiple pods, the announcing
speaker can change until all the pods are ready and
the situation is stabilized.
Adding a 5 seconds sleep should be enough for the
speakers to settle on who announces the service,
and should be enough for it to complete sending the
gratuitous ARPs as that's the time set in spamLoop().

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>