E2E: Inject the container ips from outside
Now that we are starting the containers in the setup of the suite, we
are stealing two more ips from the available ip range. This collides
with L2 tests, which are using the free ip range that is passed from
outside. By fixing the ips used by the containers, we change the
behavior to be more deterministic and we avoid having overlapping ips
between the containers and the l2 services.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>