FRR: Support eBGP multi-hop peering
Adding support for eBGP multi-hop peering in FRR is a must
as it is already supported by the native mode.

This is done by adding `ebgp-multihop` in FRR's config to a peer
that requires eBGP and is multi-hops away.

The problem with adding this without knowing if the ebgp peer is actually
multi-hops away breaks BFD, because BFD considers it multi-hop when
that parameter is set regardless if it actually is, and rejects the
peer in case it is not multi-hops away.

To solve that we introduce a new config parameter for a peer (ebgp-multihop)
that specifies if a given ebgp peer is multi-hops away or not, and setting
ebgp-multihop for that peer in FRR accordingly.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>