E2E: Change external containers setup to cover multi-hop eBGP
In order to cover multi-hop eBGP peering scenarios we need
to create another external containers that are in a different
containers network.
To achieve that we create another container network (multi-hop-net)
as part of the suite.
To have connectivity between the current containersNetwork and multi-hop-net
we create a new container that is in both, and route requests between the
networks through it by patching both the routing table of the speaker
nodes (speaker is host-networked) and the external containers that
are in the multi-hop-net.

The result is that the tests now run using 4 external FRR peers:
* 2 that are a single-hop away from the speaker - iBGP and eBGP.
* 2 multi-hops away from the speaker - iBGP and eBGP.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>