E2E tests: Fix MetalLB configmap exist check in BeforeSuite (#1126)
If we are running in operator mode the MetalLB configmap can
be found in the cluster because the operator is managing it
and not the tests. In such case BeforeSuite shouldn't fail.

This change allows us running the E2E tests more than once
in a cluster where MetalLB was deploying through operator.
With the previous validation method a second run of the tests
was failing as the configmap exist.