Check aggregation length against the wider cidr generated from range
If we specify addresses in a pool by range (and not by cidr), the check
that the lenght of the cidr is bigger than the aggregation lenght is
performed against all the possible cidrs that range translates to, and
it is misleading. The check makes sense when we declare a cidr, because
we know we own that whole cidr, but in the other case we may be using a
subset of a cidr for other reasons, and the check here leads to false
negatives.

A scenario like the one in the test that is being added should not fail:
3.3.3.2-3.3.3.254 with aggregation length 26

Here, we take the smaller (i.e. more comprehensive) netmask and we
compare the aggregation lenght to that.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>