E2E: Verify configuration in FRR containers
Add a test with a full BFD and FRR configuration.
Assertions are based on strings contained in the running configuration,
as more detailed test are already in place in internal/bgp/frr
package.

Add config.Peer.KeepaliveTime field to be aligned with the
production ConfigMap

Add substringCount ginkgo transform function to assert that a string
appears only once in configuration.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>