Mark the config as non valid if routerid is specified in FRR mode
Currently it's not possible to set different routerids per different
peers with FRR. The rendered config will contain a single routerid (the
speaker's one) per node, giving a result different from what the user
expects (a specific routerid per peer). For this reason, we remove the
configuration logic and we reject a configuration containing the
RouterID field when being applied in FRR mode.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>