Reject the configuration if multiple ASN are configured in frr mode
FRR does not allow to use different asns on the same instance (unless
using views).
The result is misleading, because the last asn added in the
configuration is used and it would be hard for users to detect.
Here we enforce the behaviour, rejecting a configuration
that sets different asn for different peers.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>