FRR: fix communities and local preferences
The current communities / local pref handling suffers of two issues:

- the route-map is added always with the same id, basically overriding
  all the communities and leaving the last one as the only available
- the route-map is not filtered per prefix. MetalLB api bounds the
  community and the local preference to the address pool, which means
that they must be a property of the announcement.

Here we make a change in the template so that every route-map entry with
the same name gets a different id, with the extra complication that we
can't bind a route-map entry to a given ip, so it's necessary to collect
the ips related to a community (or to a local pref) under a prefix-list.

E2E tests are modified to validate that only the ips coming from the
right pool are advertised with the given community / local preference.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>