FRR: Convert maps to sorted slices for prefixes before rendering
Using maps to render the template works but has one drawback. When
iterating over maps, frr does not guarantee order.
This has two major drawbacks: the unit tests with goldenfile are not
stable, but more importantly, we risk to produce different rendering
from the same config, causing un-necessary reloads.

Here we still use maps (and sets) to guarantee there are no
duplicates, but before rendering we convert the map to a sorted slice
to make the output stable.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>