test: Repeat check for frr config file
As the configuration is rendered asynchronously and
a test can produce more than one configuration, a check
against a not final configuration may occur. To avoid it,
check the file using a short poll interval.

Remove the check about output file existence, as it is now
redundant.

Add a short sleep before updating golden file to be sure
there is no config in progress.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>