L2 E2E tests: Flake fix: metrics validation of wrong speaker
In this test sometimes we get the wrong node when looking
for the advertising node of a service.
A newly created service can get the same IP the service from
the previous test got (since it was already deleted).
While the previous service is down it can take time for the
running host ARP table to be updated.

The solution for such situation is to look for the advertising
node of a service inside an eventually.

Fixing issue https://github.com/metallb/metallb/issues/1195.

An example of the failing test can be found here:
https://github.com/metallb/metallb/runs/4911555260?check_suite_focus=true