website: Document more steps for adding maintainers
This commit adds documentation for the steps we currently know how to
do them and mentions other steps that are needed but we haven't check
the specific steps to do them.

Once we go doing the other steps, we will document them here. We just
add them so we have an exhaustive list of all the things to do and don't
forget about them.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>