E2E re-order validate function
In the e2e-test bgp.go, we create external docker FRR containers
and peer them to the kind cluster nodes.
We validate that the external FRR containers are configured well
by calling the function: validateFRRPeeredWithNodes.

This commit only re-positions the calls for this function
to after calling the pairing function, to make it more
clear that it's validating the previous step in the test.