E2E: fix the create or update function
The function was mis-used. In case of update, it expects the mutatefn to
change the object. So we keep two maps (because when getting the object
the current object is modified with the version from the cluster) and we
change the object consuming the version passed to the function, using
only the spec part.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>