Helm: Move CRDs to a subchart
The recommended method in the helm documentation is
using the crds/ directory to manage the CRDs, but
this way we can't template the CRDs, which is needed
for the conversion webhooks. We decided to apply the
second recommended method and manage the CRDs in a
subchart that will be packaged with the main chart.

https://helm.sh/docs/chart_best_practices/custom_resource_definitions/#method-2-separate-charts

Changing the webhook service name is required because
it can't be templated in the subchart since helm is
adding the name of both charts to the service prefix.