Build Frr-metrics as a separate container
Move frr-metrics to a separate container instead of being
inside speaker and copied out. This makes speaker lighter and
ensures it doesn't contain code that is unwanted.

Furthermore, it helps to remove the need for speaker to have a shell,
as it no longer needs to copy frr-metrics to a shared volume.

Signed-off-by: Joshua Carnes <56089764+jtcarnes@users.noreply.github.com>