Migrate frr-reloader to its own image
Currently, frr-reloader has to have it's script copied into it using
init containers and be carried around in speaker container. This causes
speaker to need a shell to achieve the same thing as having an image
prepared.

By migrating the reloader to it's own image, it makes speaker lighter
for non-frr users, limits the use of init containers and keeps the
system locked down and secure.

Signed-off-by: Joshua Carnes <56089764+jtcarnes@users.noreply.github.com>