E2E: move the dual stack - v6 only to the multipath suite
We are trying to keep the multiprotocol tests (where we enable the
external containers to MP) isolated. The dual stack - announce v6 only
test must have the external containers enabled to mp, due to the fact
that we are announcing one single ip family (v6) but we have two open
sessions, so we are moving it.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>