CI: publish the charts after the images
The chart publishing job is faster than publishing the images, so when
we release there's a time window where the charts are ready but the
images are not, putting automatic upgrades at risk.
Here we add the chart job to the same lane of images, making it depend
on the images in order to run.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>