Remove the cidr check with ip ranges
Implementing a reliable cidr check without offering the remove buggy ips
flag is not possible.

A user that wants a full cidr without buggy ips will have to specify the
range without the extremes, and trying to reverse the logic of removing
the buggy ips is not reliable.

For this reason, we remove the consistency check between the cidr and
the aggregation length in case the cidr comes from a manual range.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>