Revert "Remove the avoidbuggyips field"
This reverts commit e4d0d18ef9fdf404c83af4ff8b43008b44fdfa9c.

We restore the avoidbuggyIP files because setting ranges is not
syntactically equivalent, if used in conjunction with aggregation
length.
A 10.0.1/24 range will become 10.0.1.1 - 10.0.1.254 which will fail a
check against the /24 aggregation length (while it should not).

Attempting to reverse the intention of the user is risky and prone to
corner cases, so it's better to restore the original configurability
option.