FRR: make the reload squashing time to 3 seconds
It normally takes 2-3 seconds to execute a full reload command. In case
of multiple requests in flight, this translates in accumulated requests
that end up in pending sighup requests that then translates into
allocated memory (basically we have one child frr_reloader process per
reload request, consuming memory).

In case a service is constantly updated (for example, a dangling
endpoint that makes the service available / not available too fast), we
may end up accumulating memory that we are not able to free fast
(because the number of requests is higher than how fast the reloader can
serve them).

Because of this, it's safer to accumulate the requests for at least how
long it takes to serve one.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit 1de7f14dbf25b0844ddaa815b2e7006da7d59d6c)