website: update backward compatibility section
This commit updates the information documented about the configmaptocrs
tool, due to the recent addition to this tool.
(Added a new way to use the tool directly on the cluster)

To keep the minimal and succinct look of the MetalLB home page,
a new page dedicated to the configmaptocrs tool was added,
and referenced from the home page.

Also updated the configmaptocrs/README.md after the change
of PR #1504 (don't default keepalive time)