Skip unnecessary node events in config reconciler
The config controller watches node event to populate l2 and bgp
advertisement objects with selected nodes based on node selector
labels.

But currently a change in the node results in reprocessing the
configuration, which results in reprocessing the services.
This results in an unnecessary cpu load.

Hence this change skips processing unnecessary node events
other than node create, node update (if its with new labels)
and node delete events.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>