Revert "When services share IP, modifying any service should fail"
This reverts commit aea1881ce3478b72d5f907655cb9a9eb32d13f72.

The meaning of the annotation is that MetalLB _may_ colocate the two
services sharing the same IP, but it's a best effort indication. If the
two services are incompatible (i.e. they have different external traffic
policies), they won't get the same IP and they should not.
This also fixes a bug where:
- the controller is stopped
- a service without an external ip is created
- a service with an external ip is already present
- the controller is started

If the controller processes the service without IP first, it may get the
first available IP (because metallb doesn't know the allocation yet),
but when the second service is processed, instead of getting a new IP
because it is not compatible with the first processed service, the
controller silently fail and the service holds the same IP as the other,
even if no request of sharing through the annotation were made.