Add task to bump image in values file and workflow to bump image in main
This adds a task that can be used during the release flow to ensure that
the values.yaml file contains the latest release. It also adds a seperate
workflow that only runs on a push to main that will bump the values.yaml
file to use the latest release which will allow users to deploy the helm
chart directly from the main branch.

Signed-off-by: Tyler Auerbeck <tylerauerbeck@users.noreply.github.com>