Refactor the frr configuration template
Removing duplication and isolating separate parts in separate
subtemplates to make it a bit easier to navigate it.

We also change the way the prefixes are being added, because the
previous version was misleading in how the configuration was rendered,
making the reader think that the prefixes are advertised per neighbor
while they are per router (and we have per neighbor specific filters to
choose what prefixes should go to a given neighbor).

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit 1271c56243f31d5fbffa056d2eca647cd2d64e19)