Refactor: change newSession to accept a structure
The newSession method parameter list is growing and it is hard to keep
track of it. Here we change the function to accept a struct with the
list of arguments, and we embed the structure in the session object used
by both implementations.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit c368680e117fed7976de042efd4b6c571b27de97)