FRR templates: make the order deterministic
We have a few config fields (or subfields) that are maps, which makes
the rendering non deterministic because ranging over a map doesn't give
guarantee over the order. So, we decouple the internal representation of
the config from the one we feed the template with, and we feed the
template with slices instead of maps.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>