L2: continue if some responders cannot be created
If a responder cannot be created on a given interface, it should not
prevent responders to be created on all the other interfaces. For this
reason here we continue instead of returning. This is even more sensible
with the interface selector mechanism, because returning won't allow the
list of interfaces to be filled and so the speaker will always skip.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>