E2E: use vrf containers only in frr mode
Native mode does not support VRFs, so we can't use the vrf containers
when running on the native lane.

Because of this, we add a parameter that drives the set of containers we
want to use as base for the tests.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit 84622fa42e17b5d4e970a896dc008a3ef2d78845)