FRR liveness probe: move the execution inside the container
Using an exec probe means not binding the resource consumption of the
execution of the probe to the container, but to the container runtime.

By moving the execution of the vtysh behind the http endpoint, we ensure
that all the cpu quotas are respected because the actual vtysh execution
happens in the context of the container.

Despite the liveness probe is related to the frr container, we execute
it from the frr-metrics container which has already access to vtysh.
This, in order not to pollute the frr container with an additional
process, and it is possible since all the containers in the pod share
the same network namespace.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit ae831ef9d6cf5d6dc344f0fbd5907301379c46b8)