Assignement tests: make the generic pool broader
If there is a service of type loadbalancer in the cluster we are testing
against, it will steal the single ip we expect to be assigned to our
services. For this reason, we broaden the number of ips available for
the default svc in order to have free ips available for our test svc.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit b31e1bd58ba0e174d7682159dae2e5691b1b5773)