E2E: don't exit if not able to find multihopRoutes
When running on host mode, the nexthop container is not running so
getting the routes will fail. So, we check if the the routes are not
fetched because the container is not there, and in case we continue and
not error.

This fixes the error on aftersuite when running the tests in host mode.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>