Config Controller: ignore changes that are not relevant
The configuration controller is triggered by multiple events, including
all the secrets in the metallb's namespace. This cause a cascade effect
of reprocessing all the services which is not necessary.

Here we compare the last processed configuration with the one we just
produced (which takes into account of non used secrets and non used
namespaces, for example), limiting configuration reload events to those
that really are changing the metallb's internal configuration.

On top of that, in order to keep the node label test effective, the
implementation of the reconciliation function is deferred to a local
function that can be overridden by the tests.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit fd4ca5ed91d184e3bdf74797552473fa8994bf1f)