service_controller_test: add an indexer function
The new implementation of mock client requires to be fed with the right
indexer function in order to retrieve the right objects.

Here we add the real function and we add the label used by the indexer.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit 6daec72c5e48a840e251bf1a188352f1585213fc)