Make the exit more robust
When we clear, we need to understand if the multihop container is
running, because otherwise we'll panic in inspecting it (and in removing
the multihop wiring). Checking the error depends on the container
runtime, so it's a poor choice. Here we check if the container is among
those we are processing, and if not we don't tear down the multihop
setup.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit fb9fb96353b010e16d32462d1ae37e8536915ddf)