Make localPref uniform among updates
These changes ensure that the local preference is unique among BGP ADVs
with the same type of BGP update. It is still possible to have
BGP ADVs with different values of localPref if the set of BGP peers are different.
BGP ADVs with common BGP peers must not have common nodes. Checks are performed
for BGP ADVs that have the same aggregation length.

Fixed #1739

Signed-off-by: Marcelo Guerrero Viveros <marguerr@redhat.com>