frr template: add the deny any prefixlist only when no advs
The frr doc says "If no ip prefix-list is specified, it acts
as permit. If ip prefix-list is defined, and no match is
found, default deny is applied"

Here we add a deny any only if there are no advertisements, to handle
special scenarios where we'd have a prefix but not advertised to a
specific peer because of the peer selector.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>