E2ETests: remove the dependencies from metallb internal
We make the e2etests a standalone go module, with no dependencies from
the metallb's internal packages.
This is done with a small amount of repetition (the packages that were
consumed directly by metallb/internal are now copied under e2etest/pkg),
but this is for a greater scope which is to make the e2e tests a
standalone module dependant only from the MetalLB API.

This will allow to reuse the infrastructure framework that was created
in the testsuite also for the FRR Daemon, and at some point to move the
e2e framework in a separated repo consumed by both.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>