Add go.work to the e2etests
The e2etests are meant to consume the local version of MetalLB. At the
same time, they are (currently) exported as a versioned framework for
implementing tests.

Since they belong to a separate module now, the way they'd work is to
pull the latest tagged version of the MetalLB API, which makes the
development not practical. By adding a local go.work file, we tell the
e2e to consume the local version of the MetalLB module (basically, the
API which is the only remained dependency) so we can evolve MetalLB and
the tests together without incurring in chicken - egg issues.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>