Remove IP verifications in SetPools
Pool changes might invalidate the LB IPs of services. When this
happens, the pool controller ignores the new configuration and
continues operating with the current one. This PR modifies this
behavior by always accepting the new configuration and letting the
service controller to clear the IPs of the affected services.

Fixes #462

Signed-off-by: Marcelo Guerrero Viveros <marguerr@redhat.com>
Co-authored-by: Federico Paolinelli <fpaoline@redhat.com>