Make the configuration conversion stable
When translating the configuration coming from the k8s api to the
internal one, the outcome depends on the order in which the cache
returns us the various items. This may cause unnecessary reloads.

In order to avoid this, we pre-order the list of items to feed the
conversion with.
The sorting is done in a conversion function in the controller
(rather than the config package) in order to make it harder to
forget to add the sorting action.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>