Configmaptocrs now converts resource names
Previously, if a resource such as a BFDProfile or IpAddressPool
had invalid names, the generated YAML files would not be valid.
This would happen if they:
1. Did not start with a letter
2. Had uppercase characters
3. Had underscores
4. Were over 63 characters long

Not automatically converting the names would introduce problems
in automated workflows.
A warning appears in the configmaptocrs container logs for each
converted resource name

Signed-off-by: Sebastian-RG <srojas@whitestack.com>