Add an FRRK8sConfig controller
The controller is driven in two ways:
- it listens for changes on the frrconfig related to the metallb
  namespace and the node the controller is running on
- it listen for changes in the UpdateConfig method

Given calling UpdateConfig is asynchronous, the passed data is stored
and an event is sent to a channel so it can be processed inside
the reconciliation loop.

The notification callback has empty interface as a field so this
callback an be used as a bgp event listener.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>