E2E: adjust tests so we can run them on frr-k8s mode
Using frr-k8s mode we can no longer depend on frr being a sidecar of the
speaker. We introduce a new "FRRProvider" interface that allows the code
to fetch the executor matching the frr container seamlessly without caring what mode
it is running on, by setting bgptests.FRRProvider on suite startup according to
the given mode.

In addition, the interface allows fetching the bgp metrics pod and the prefix (prometheus namespace)
to check against, since frr-k8s and metallb expose the same metrics under a different prefix.

Lastly, the metrics scraping is adjusted to work against the https endpoints instead of http
as frr-k8s doesn't expose them - this will also enable us to not listen externally on the
http endpoints in MetalLB like we do in frr-k8s.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>