BGP: move the session matching logic one level above
Sending all the advertisements to the session just to filter them after
is counterintuitive. One would expect all of them to be processed. It's
much better (and readable) to filter them from the calling site, before
passing them to the session.

Also, removing all the *selector* tests from frr and frr-k8s since they
do not make sense as the logic is happening above.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>