FRR: squash advertisements with the same prefix
Passing multiple advertisements with the same prefix to the template
causes duplicate entries. This happens for example when the shared ip
annotation is used with the services.

Since a given advertisement has the related peers as fields, the right
place to squash is when we manage the advertisements related to a given
neighbor.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>