Tasks bumprelease: bump only the version of the crds subchart
The regex was now matching also the version of frr-k8s, which is fixed
and must be reflected in an explicit change in the codebase.

Here we only change if the previous version is crds.enabled, meaning we
are referring to the crds section.

Also, align all the other helm replacements to use sed instead of perl.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>