E2E:  validate FRR running configuration when connect time included
This test ensures that the  connectTime setting specified in the
MetalLB BGPeer CRD is reflected in the FRR running configuration. The
test uses the frr.NeighborsInfo function to get the neighbors info
in a parsed form. The function returns the Neighbor struct that is updated to include the ConfiguredConnectTime field.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>