CI: trigger the add labels job on pull_request_target
We don't want to allow the labeler job to change ANY PR, just the one
the action is currently running against.

From the docs:
This event runs in the context of the base of the pull request,
rather than in the context of the merge commit, as the pull_request
event does.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>