E2E:  validate FRR running config when MP BGP is disabled
Added e2e test that ensures that the disableMP setting
specified in the MetalLB BGPeer CRD is reflected in the FRR running configuration. The test uses the frr.NeighborsInfo function to get the neighbors info in a parsed form. The function returns the Neighbor struct that is updated to include a list of the names of the address families that are used by the test to ensure that ipv4 addresses are not set in the ipv6 family and viceversa.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>