Speaker: hide the L2ServiceStatus generation behind a flag
Because of https://github.com/metallb/metallb/issues/2311 we are going
to move the status instances to the metallb namespace. This might
require a change in the permissions too (from cluster to namespaced) so
the new version of MetalLB might not be able to delete the "legacy"
instances of the CRD because they belong to namespaces the new metallb
might not have permissions on.

Because of this, we hide the feature behind a flag, effectively
disabling it until the issue is fixed.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>