Speaker Chart: listen to localhost if securemetricsport is present
When we instantiate Values.prometheus.secureMetricsPort, we add a
sidecar kuberbac proxy with the metrics port as upstream.

Because of that, it's fair not to expose the unencrypted metrics port
from the node itself, as it's exposed to the world.

Here we also change the metrics exporter to be able to listen to
localhost only.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>