frr-k8s: Add disableMP field
Added BGP disableMP field in the frrk8s 
BGP session manager. This change enables 
the setting of disableMP in frrk8s mode. 
This allowed to enable  the test run of "BGP peer disable MP BGP" to run in frrk8s mode. To remove code duplication the "BGP peer disableMP" was combined with the  "BGP peer connect time" test into one test that checks the setting of parameters in BGP peer.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>