API: add gracefulRestart support
BGP graceful restart functionality as defined in RFC-4724 defines the
mechanisms that allows BGP speaker to continue to forward data packets
along known routes while the routing protocol information is being
restored. This allows DS to be updated without routes being retracted in
the peer side.

We enable by default if GR then the F-bit (preserve-fw-state) to be set.

We make gracefulRestart immutable according to
https://kubernetes.io/blog/2022/09/29/enforce-immutability-using-cel/#immutablility-after-first-modification

Signed-off-by: karampok <karampok@gmail.com>