don't log history in liveness probe
frr-metrics liveness probe runs a show daemons command repeatedly and
append that at the end of a file. When the file is longer than 1k lines
the history gets truncated which involves a few system calls like
renames. When the system is under stress this rename call can take a few
seconds, exceeding the timeout of the liveness probe.

When this happens repeatedly frr process gets restarted.

This was noticed as part of the investigation in
https://issues.redhat.com/browse/OCPBUGS-34761?focusedId=24850879&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-24850879

Signed-off-by: François Rigault <frigo@amadeus.com>