feat: Add the speaker owner reference to the l2 service status instance
In order to avoid dangling status resources, we add an owner reference from a speaker to the L2Service status it owns.

Because of that, we also move the instance to the MetalLB's namespace.

Labels are added to make it easier to understand which node / speaker is announcing a given service.

Signed-off-by: lwabish <wubw@pku.edu.cn>