Support running frr-k8s in a separate namespace
We allow frr-k8s to run in a separate namespace. If that is the case,
the password coming from secrets in the metallb's namespace is
translated to plain text in the frrconfiguration.

This is to avoid giving metallb permissions to cluster-scoped secrets,
and acceptable as from a security point of view, having access to the
cluster gives the same level of security between secrets and
frrconfigurations.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>