Service annotations: change annotation prefix to metallb.io
This commit deprecates the old annotation prefix metallb.universe.tf in
favor of metallb.io.

The introduced change is backwards compatible. E.g., if a services
specifies a address-pool using the old prefix, it will still be accepted
by the controller. However, if two address-pool annotations are
specified the one with the newer prefix (metallb.io/address-pool) has
precedence.

If a services uses the old prefix an warning event for the corresponding
service is generated. Moreover, a log entry will be generated (severity
warning).

Fixed #2392

Signed-off-by: Adrian Vondendriesch <adrian.vondendriesch@netapp.com>