Define default holdtime value in CRD
If we have holdtime default value defined in code and not in the CR then
the `kubectl get bgpeer -o yaml` does not show the real value, which is
not ideal.

We add validation directly in the CRD
(holdtime >3s, keepalivetime more than 1s)

We extend the keepalivetime validation in webhook to check if the value
is less than 1/3 of the holdtime. If keepalivetime is larger then FRR
just change it and this is not visible unless

```
k -n metallb-system exec -it speaker-59rlv – /bin/bash
vtysh; show running-config` inside a speaker pod

you will see also that finally the entry appears to be
  neighbor 192.168.111.1 timers 30 90
```

Signed-off-by: karampok <karampok@gmail.com>